Move code from LifecycleManager to BaseLifecycleManager
